requirements.txt
annotated-types==0.7.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.30.2
pvcobra==2.0.2
pvporcupine==3.0.2
PyAudio==0.2.14
pydantic==2.7.1
pydantic_core==2.18.2
pygame==2.5.2
python-decouple==3.8
requests==2.32.2
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
pydub==0.25.1
numpy==1.24.4
python-decouple==3.8

